+++ lib/map/map_world.dart
@@ -13,13 +13,18 @@ class MapWorld extends MapGame {
   Vector2 lastCamera = Vector2.zero();
   double lastMinorZoom = 1.0;
   Vector2? lastSizeScreen;
-  List<ObjectCollision> _tilesCollisions = List.empty();
+  List<ObjectCollision> _tilesCollisions = List.empty(growable: true);
   List<ObjectCollision> _tilesVisibleCollisions = List.empty();
   List<TileModel> _tilesToAdd = [];
   List<Tile> _tilesToRemove = [];
   Set<String> _visibleSet = Set();
   bool _buildingTiles = false;
+  bool _updateMapSize = true;
+  bool _updateStartPosition = true;
   double tileSize = 0.0;
+  Vector2 _griSize = Vector2.zero();
+  Size _mapSize = Size.zero;
+  Vector2 _mapStartPosition = Vector2.zero();
 
   List<Offset> _linePath = [];
   Paint _paintPath = Paint()
@@ -51,7 +56,7 @@ class MapWorld extends MapGame {
   void update(double dt) {
     for (Tile tile in childrenTiles) {
       tile.update(dt);
-      if (tile.shouldRemove) {
+      if (tile.isRemoving) {
         _tilesToRemove.add(tile);
       }
     }
@@ -105,24 +110,30 @@ class MapWorld extends MapGame {
   @override
   void onGameResize(Vector2 size) {
     if (isLoaded) {
-      _verifyMaxTopAndLeft(size);
+      _createQuadTree(size);
     }
     super.onGameResize(size);
   }
 
-  void _verifyMaxTopAndLeft(Vector2 size, {bool isUpdate = false}) {
+  void _createQuadTree(Vector2 size, {bool isUpdate = false}) {
     if (lastSizeScreen == size) return;
     lastSizeScreen = size.clone();
 
     if (isUpdate) {
       lastCamera = Vector2.zero();
       lastMinorZoom = gameRef.camera.zoom;
+      _updateMapSize = true;
+      _updateStartPosition = true;
     }
 
     tileSize = tiles.first.width;
 
-    mapSize = getMapSize();
-    mapStartPosition = getStartPosition();
+    final mapSize = getMapSize();
+
+    _griSize = Vector2(
+      (mapSize.width.ceil() / tileSize).ceilToDouble(),
+      (mapSize.height.ceil() / tileSize).ceilToDouble(),
+    );
 
     if (tileSizeToUpdate == 0) {
       tileSizeToUpdate = (tileSize * 4).ceilToDouble();
@@ -138,8 +149,8 @@ class MapWorld extends MapGame {
       quadTree = QuadTree(
         0,
         0,
-        ((mapSize?.width ?? 0).ceil() / tileSize).ceil(),
-        ((mapSize?.height ?? 0).ceil() / tileSize).ceil(),
+        _griSize.x,
+        _griSize.y,
         maxItems: maxItems,
       );
 
@@ -154,24 +165,31 @@ class MapWorld extends MapGame {
     lastSizeScreen = null;
     this.tiles = map;
     await Future.forEach<TileModel>(tiles, _loadTile);
-    _verifyMaxTopAndLeft(gameRef.size, isUpdate: true);
+    _createQuadTree(gameRef.size, isUpdate: true);
   }
 
   @override
   Size getMapSize() {
-    double height = 0;
-    double width = 0;
+    if (_updateMapSize && tiles.isNotEmpty) {
+      double height = 0;
+      double width = 0;
 
-    this.tiles.forEach((tile) {
-      if (tile.right > width) width = tile.right;
-      if (tile.bottom > height) height = tile.bottom;
-    });
+      this.tiles.forEach((tile) {
+        if (tile.right > width) width = tile.right;
+        if (tile.bottom > height) height = tile.bottom;
+      });
+      _updateMapSize = false;
+      return _mapSize = Size(width, height);
+    }
 
-    return Size(width, height);
+    return _mapSize;
   }
 
+  Vector2 getGridSize() => _griSize;
+
+  @override
   Vector2 getStartPosition() {
-    try {
+    if (_updateStartPosition && this.tiles.isNotEmpty) {
       double x = this.tiles.first.left;
       double y = this.tiles.first.top;
 
@@ -179,10 +197,10 @@ class MapWorld extends MapGame {
         if (tile.left < x) x = tile.left;
         if (tile.top < y) y = tile.top;
       });
-
-      return Vector2(x, y);
-    } catch (e) {
-      return Vector2.zero();
+      _updateStartPosition = false;
+      return _mapStartPosition = Vector2(x, y);
+    } else {
+      return _mapStartPosition;
     }
   }
 
@@ -206,16 +224,15 @@ class MapWorld extends MapGame {
   }
 
   void _getTileCollisions() {
-    List<ObjectCollision> aux = [];
+    _tilesCollisions.clear();
     final list = tiles.where((element) {
       return element.collisions?.isNotEmpty == true;
     });
 
     for (var element in list) {
-      final o = element.getTile(gameRef);
-      aux.add(o as ObjectCollision);
+      final collision = element.getTile(gameRef);
+      _tilesCollisions.add(collision as ObjectCollision);
     }
-    _tilesCollisions = aux;
   }
 
   List<Tile> _buildTiles(Iterable<TileModel> visibleTiles) {
@@ -228,14 +245,14 @@ class MapWorld extends MapGame {
   Future<void>? onLoad() async {
     await super.onLoad();
     await Future.forEach<TileModel>(tiles, _loadTile);
-    _verifyMaxTopAndLeft(gameRef.size);
+    _createQuadTree(gameRef.size);
     _searchTilesToRender();
   }
 
   void _verifyRemoveTileOfWord() {
     if (_tilesToRemove.isNotEmpty) {
       for (Tile tile in _tilesToRemove) {
-        if (tile.shouldRemove) {
+        if (tile.isRemoving) {
           childrenTiles.remove(tile);
           tiles.removeWhere((element) => element.id == tile.id);
           quadTree?.removeById(tile.id);
@@ -270,6 +287,9 @@ class MapWorld extends MapGame {
       _tilesCollisions.add(tile as ObjectCollision);
       _findVisibleCollisions();
     }
+
+    _updateMapSize = true;
+    _updateStartPosition = true;
   }
 
   @override
@@ -281,6 +301,8 @@ class MapWorld extends MapGame {
     } catch (e) {
       print('Not found visible tile with $id id');
     }
+    _updateMapSize = true;
+    _updateStartPosition = true;
   }
 
   void _findVisibleCollisions() {
@@ -300,13 +322,6 @@ class MapWorld extends MapGame {
     return Future.value();
   }
 
-  Vector2 _getCameraTileUpdate() {
-    return Vector2(
-      (gameRef.camera.position.x / tileSizeToUpdate).floorToDouble(),
-      (gameRef.camera.position.y / tileSizeToUpdate).floorToDouble(),
-    );
-  }
-
   bool _checkNeedUpdateTiles() {
     final camera = _getCameraTileUpdate();
     if (lastCamera != camera || lastMinorZoom > gameRef.camera.zoom) {
@@ -317,5 +332,12 @@ class MapWorld extends MapGame {
       return true;
     }
     return false;
+  }
+
+  Vector2 _getCameraTileUpdate() {
+    return Vector2(
+      (gameRef.camera.position.x / tileSizeToUpdate).floorToDouble(),
+      (gameRef.camera.position.y / tileSizeToUpdate).floorToDouble(),
+    );
   }
 }